requirements.txt
pycups
-e git://github.com/twam/pyzpl2.git#egg=pyzpl2

